requirements.txt
beautifulsoup4==4.11.2
python-dotenv==1.0.0
Requests==2.31.0
selenium==4.16.0
webdriver_manager==4.0.1
